Move consecutive collector out of examples
