+++ core/core-impl/src/main/java/ai/timefold/solver/core/api/score/stream/impl/SequenceImpl.java
@@ -1,12 +1,12 @@
-package ai.timefold.solver.examples.common.experimental.impl;
+package ai.timefold.solver.core.api.score.stream.impl;
 
 import java.util.Iterator;
 import java.util.NavigableSet;
 import java.util.function.BiFunction;
 import java.util.stream.Collectors;
 
-import ai.timefold.solver.examples.common.experimental.api.Break;
-import ai.timefold.solver.examples.common.experimental.api.Sequence;
+import ai.timefold.solver.core.api.score.stream.api.Break;
+import ai.timefold.solver.core.api.score.stream.api.Sequence;
 
 final class SequenceImpl<Value_, Point_ extends Comparable<Point_>, Difference_ extends Comparable<Difference_>>
         implements Sequence<Value_, Difference_> {
