+++ examples/src/test/java/ai/timefold/solver/examples/common/experimental/ExperimentalConstraintCollectorsTest.java
@@ -3,83 +3,21 @@ package ai.timefold.solver.examples.common.experimental;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Arrays;
-import java.util.Objects;
 import java.util.function.BiConsumer;
 
 import ai.timefold.solver.core.api.score.stream.uni.UniConstraintCollector;
-import ai.timefold.solver.examples.common.experimental.api.ConsecutiveInfo;
 import ai.timefold.solver.examples.common.experimental.impl.ConsecutiveIntervalInfoImpl;
-import ai.timefold.solver.examples.common.experimental.impl.ConsecutiveSetTree;
 import ai.timefold.solver.examples.common.experimental.impl.IntervalTree;
 
 import org.junit.jupiter.api.Test;
 
 class ExperimentalConstraintCollectorsTest {
-    @Test
-    void consecutive() {
-        // Do a basic test w/o edge cases; edge cases are covered in ConsecutiveSetTreeTest
-        var collector = ExperimentalConstraintCollectors.consecutive(Integer::intValue);
-        var container = collector.supplier().get();
-        // Add first value, sequence is [2]
-        int firstValue = 2;
-        Runnable firstRetractor = accumulate(collector, container, firstValue);
-        assertResult(collector, container, consecutiveData(2));
-        // Add second value, sequence is [1,2]
-        int secondValue = 1;
-        Runnable secondRetractor = accumulate(collector, container, secondValue);
-        assertResult(collector, container, consecutiveData(1, 2));
-        // Add third value, same as the second. Sequence is [{1,1},2}]
-        Runnable thirdRetractor = accumulate(collector, container, secondValue);
-        assertResult(collector, container, consecutiveData(1, 1, 2));
-        // Retract one instance of the second value; we only have two values now.
-        secondRetractor.run();
-        assertResult(collector, container, consecutiveData(1, 2));
-        // Retract final instance of the second value; we only have one value now.
-        thirdRetractor.run();
-        assertResult(collector, container, consecutiveData(2));
-        // Retract last value; there are no values now.
-        firstRetractor.run();
-        assertResult(collector, container, consecutiveData());
-    }
-
-    private static class Interval {
-        final int start;
-        final int end;
-
-        public Interval(int start, int end) {
-            this.start = start;
-            this.end = end;
-        }
-
-        public int getStart() {
-            return start;
-        }
-
-        public int getEnd() {
-            return end;
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o)
-                return true;
-            if (o == null || getClass() != o.getClass())
-                return false;
-            Interval interval = (Interval) o;
-            return start == interval.start && end == interval.end;
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(start, end);
-        }
-    }
 
     @Test
     void consecutiveInterval() {
         // Do a basic test w/o edge cases; edge cases are covered in ConsecutiveSetTreeTest
         var collector =
-                ExperimentalConstraintCollectors.consecutiveIntervals(Interval::getStart, Interval::getEnd, (a, b) -> b - a);
+                ExperimentalConstraintCollectors.consecutiveIntervals(Interval::start, Interval::end, (a, b) -> b - a);
         var container = collector.supplier().get();
         // Add first value, sequence is [(1,3)]
         Interval firstValue = new Interval(1, 3);
@@ -103,16 +41,9 @@ class ExperimentalConstraintCollectorsTest {
         assertResult(collector, container, consecutiveIntervalData());
     }
 
-    private ConsecutiveInfo<Integer, Integer> consecutiveData(Integer... data) {
-        return Arrays.stream(data).collect(
-                () -> new ConsecutiveSetTree<Integer, Integer, Integer>((a, b) -> b - a, Integer::sum, 1, 0),
-                (tree, datum) -> tree.add(datum, datum),
-                mergingNotSupported());
-    }
-
     private ConsecutiveIntervalInfoImpl<Interval, Integer, Integer> consecutiveIntervalData(Interval... data) {
         return Arrays.stream(data).collect(
-                () -> new IntervalTree<>(Interval::getStart, Interval::getEnd, (a, b) -> b - a),
+                () -> new IntervalTree<>(Interval::start, Interval::end, (a, b) -> b - a),
                 (tree, datum) -> tree.add(tree.getInterval(datum)),
                 mergingNotSupported()).getConsecutiveIntervalData();
     }
@@ -137,4 +68,9 @@ class ExperimentalConstraintCollectorsTest {
                 .ignoringFields("sourceTree", "indexFunction", "sequenceList", "startItemToSequence")
                 .isEqualTo(expectedResult);
     }
+
+    private record Interval(int start, int end) {
+
+    }
+
 }
